testish added

+++ src/Clock.java
@@ -14,7 +14,7 @@ public class Clock extends Meter {
         return accuracy;
     }
 
-    public void setAccuracy(float accuracy) {
+    public void setAccuracy(double accuracy) {
         this.accuracy = accuracy;
     }
